deliberative_compute/requirements.txt
transformers>=4.42.0
accelerate>=0.33.0
datasets>=2.20.0
evaluate>=0.4.2
sentencepiece
tqdm
numpy
pyyaml
matplotlib
jsonlines
regex
scikit-learn
psutil
# hosted APIs
openai>=1.14.3
google-generativeai>=0.8.3
# energy (optional)
pynvml
